QSDK:PROTOCOL: ath12k: add vendor regdomain param and cache FW regdomain state
Add `QCA_WLAN_VENDOR_RADIO_PARAM_REGDOMAIN` and store the active
regdomain in `ath12k_radio_cfg`.

Update regulatory event handling to build and cache a packed regdomain
value from firmware channel list updates: lower 16 bits for the 2G/5G
regdomain pair and upper 16 bits for the 6 GHz super-domain. Include the
6 GHz super-domain bits only when 6 GHz AP rules are present.

Also propagate alpha2 and country_id from the same update to all radios
to keep per-radio regulatory state synchronized.

Change `ath12k_wmi_init_country_arg.cc_info.regdom_id` to `u32` to match
the packed regdomain representation.

Sample command:
cfg80211tool phyX radio_idx Y set_regdomain <value>
cfg80211tool phyX radio_idx Y get_regdomain

Tag: REGULATORY

Patch-dependency: None
Patch-work: None

+++ backports-6.1-9a0dddfb3/drivers/net/wireless/ath/ath12k/wmi.h
@@ -5487,7 +5487,7 @@ struct wmi_therm_throt_level_config_info {
 struct ath12k_wmi_init_country_arg {
 	union {
 		u16 country_code;
-		u16 regdom_id;
+		u32 regdom_id;
 		u8 alpha2[3];
 	} cc_info;
 	enum set_init_cc_flags flags;
